LEDGE-161: cleanups: redunant imports, superinterfaces, declared exceptions, casts

+++ ledge-web/src/main/java/org/objectledge/web/mvc/tools/EmailTool.java
@@ -69,6 +69,6 @@ public class EmailTool
      */
     private String bin2Hex(char c)
     {
-        return Integer.toHexString((int)c);
+        return Integer.toHexString(c);
     }
 }
